% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_year_2016_to_ACT_ppr}
\alias{add_year_2016_to_ACT_ppr}
\title{Add year 2016 to ACT population projection}
\usage{
add_year_2016_to_ACT_ppr(
  aus_sa3_ACT_ppr_2018_tb,
  aus_sa2_ACT_erp_age_sex_2016_tb
)
}
\arguments{
\item{aus_sa3_ACT_ppr_2018_tb}{Australia Statistical Area 3 ACT population projection 2018 (a tibble)}

\item{aus_sa2_ACT_erp_age_sex_2016_tb}{Australia Statistical Area 2 ACT Estimatedesident Population age sex 2016 (a tibble)}
}
\description{
add_year_2016_to_ACT_ppr() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add year 2016 to act population projection. Function argument aus_sa3_ACT_ppr_2018_tb specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
