% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aus-package.R, R/pkg_aus.R
\docType{package}
\name{aus-package}
\alias{aus}
\alias{aus-package}
\title{aus: Model Australian Spatial Data with Ready4}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for modelling Australia specific geometry and spatial attribute data for use with the ready4 youth mental health systems model (https://ready4-dev.github.io/ready4/). This development version of the aus package has been made available as part of the process of testing and documenting the package. It is currently highly unstable and is not yet recommended for use. If you have any questions, please contact the authors (matthew.hamilton@orygen.org.au).

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for modelling Australia specific geometry and spatial attribute data for use with the ready4 youth mental health systems model (https://ready4-dev.github.io/ready4/). This development version of the aus package has been made available as part of the process of testing and documenting the package. It is currently highly unstable and is not yet recommended for use. If you have any questions, please contact the authors (matthew.hamilton@orygen.org.au).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/aus/}
  \item \url{https://github.com/ready4-dev/aus}
  \item \url{https://ready4-dev.github.io/ready4/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/aus/}
  \item \url{https://github.com/ready4-dev/aus}
  \item \url{https://ready4-dev.github.io/ready4/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Headspace [funder]
  \item National Health and Medical Research Council [funder]
}

}
\keyword{internal}
