% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_author.R
\name{author-AusACT}
\alias{author-AusACT}
\alias{author,AusACT-method}
\alias{author-AusHeadspace}
\alias{author,AusHeadspace-method}
\alias{author-AusProjections}
\alias{author,AusProjections-method}
\alias{author-AusOrygen}
\alias{author,AusOrygen-method}
\title{Author and save files}
\usage{
\S4method{author}{AusACT}(x, what_1L_chr = "ppr_2018")

\S4method{author}{AusHeadspace}(x)

\S4method{author}{AusProjections}(x)

\S4method{author}{AusOrygen}(x)
}
\arguments{
\item{x}{An object of class AusOrygen}

\item{what_1L_chr}{What (a character vector of length one), Default: 'ppr_2018'}
}
\value{
Y (Look up tables for spatiotemporal data)

Y (Look up tables for spatiotemporal data)

Y (Look up tables for spatiotemporal data)

Y (Look up tables for spatiotemporal data)
}
\description{
author method applied to AusACT

author method applied to AusHeadspace

author method applied to AusProjections

author method applied to AusOrygen
}
