% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decis.R
\name{change_final_decis}
\alias{change_final_decis}
\title{change_final_decis <- function(x) {
ifelse(
grepl("D0", x) | jaid::count_pattern("E0", x) >= 4, "D0", ifelse(
grepl("NR|U0", x), "U0", ifelse(
grepl("RA|A0|E0", x), gsub("S0|,S0$", "", x), x)
)
)
}}
\usage{
change_final_decis(x)
}
\description{
change_final_decis <- function(x) {
ifelse(
grepl("D0", x) | jaid::count_pattern("E0", x) >= 4, "D0", ifelse(
grepl("NR|U0", x), "U0", ifelse(
grepl("RA|A0|E0", x), gsub("S0|,S0$", "", x), x)
)
)
}
}
