% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_no_stat_areas}
\alias{get_no_stat_areas}
\title{Get no statistic areas}
\usage{
get_no_stat_areas(
  land_sf,
  statistic_sf,
  land_name_1L_chr,
  land_val_vec,
  stat_name_str,
  stat_val_vec
)
}
\arguments{
\item{land_sf}{Land (a simple features object)}

\item{statistic_sf}{Statistic (a simple features object)}

\item{land_name_1L_chr}{Land name (a character vector of length one)}

\item{land_val_vec}{PARAM_DESCRIPTION}

\item{stat_name_str}{PARAM_DESCRIPTION}

\item{stat_val_vec}{PARAM_DESCRIPTION}
}
\description{
get_no_stat_areas() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get no statistic areas. Function argument land_sf specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
