% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_abbreviations_lup}
\alias{make_abbreviations_lup}
\title{Make abbreviations lookup table}
\usage{
make_abbreviations_lup()
}
\value{
Abbreviations (a ready4 S3 extension of lookup table)
}
\description{
make_abbreviations_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make abbreviations lookup table. The function returns Abbreviations (a ready4 S3 extension of lookup table).
}
\keyword{internal}
