% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_correspondences_lup}
\alias{make_correspondences_lup}
\title{Make correspondences lookup table}
\usage{
make_correspondences_lup()
}
\description{
make_correspondences_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make correspondences lookup table. The function is called for its side effects and does not return a value.
}
\keyword{internal}
