% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_geom_imports_lup}
\alias{make_geom_imports_lup}
\title{Make geometry imports lookup table}
\usage{
make_geom_imports_lup()
}
\description{
make_geom_imports_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make geometry imports lookup table. The function is called for its side effects and does not return a value.
}
\keyword{internal}
