% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_no_stat_ls}
\alias{make_no_stat_ls}
\title{Make no statistic list}
\usage{
make_no_stat_ls(
  area_atts_ls,
  area_geoms_ls,
  feature_pfx_1L_chr,
  land_sf,
  land_name_1L_chr,
  statistic_sfs_ls,
  uids_chr,
  var_nms_chr
)
}
\arguments{
\item{area_atts_ls}{Area attributes (a list)}

\item{area_geoms_ls}{Area geometries (a list)}

\item{feature_pfx_1L_chr}{Feature prefix (a character vector of length one)}

\item{land_sf}{Land (a simple features object)}

\item{land_name_1L_chr}{Land name (a character vector of length one)}

\item{statistic_sfs_ls}{Statistic simple features objects (a list)}

\item{uids_chr}{Unique identifiers (a character vector)}

\item{var_nms_chr}{Variable names (a character vector)}
}
\description{
make_no_stat_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make no statistic list. The function is called for its side effects and does not return a value.
}
\keyword{internal}
