% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_spine_no_stat_ls}
\alias{make_spine_no_stat_ls}
\title{Make spine no statistic list}
\usage{
make_spine_no_stat_ls(
  area_atts_chr,
  area_geoms_chr,
  land_sf,
  statistic_sf,
  land_name_1L_chr,
  stat_name_str
)
}
\arguments{
\item{area_atts_chr}{Area attributes (a character vector)}

\item{area_geoms_chr}{Area geometries (a character vector)}

\item{land_sf}{Land (a simple features object)}

\item{statistic_sf}{Statistic (a simple features object)}

\item{land_name_1L_chr}{Land name (a character vector of length one)}

\item{stat_name_str}{PARAM_DESCRIPTION}
}
\description{
make_spine_no_stat_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make spine no statistic list. The function is called for its side effects and does not return a value.
}
\keyword{internal}
