% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture-AusLookup}
\alias{manufacture-AusLookup}
\alias{manufacture,AusLookup-method}
\title{Manufacture a new object}
\usage{
\S4method{manufacture}{AusLookup}(
  x,
  area_sf = NULL,
  att_data_xx = NULL,
  area_unit_1L_chr = character(0),
  include_chr = character(0),
  match_value_xx = character(0),
  population_tb = NULL,
  region_short_long_chr = character(0),
  subdivision_1L_chr = character(0),
  type_1L_chr = "custom",
  what_1L_chr = "attribute",
  year_1L_chr = character(0),
  ...
)
}
\arguments{
\item{x}{An object of class AusLookup}

\item{area_sf}{Area (a simple features object), Default: NULL}

\item{att_data_xx}{Attribute data (an output object of multiple potential types), Default: NULL}

\item{area_unit_1L_chr}{Area unit (a character vector of length one), Default: character(0)}

\item{include_chr}{Include (a character vector), Default: character(0)}

\item{match_value_xx}{Match value (an output object of multiple potential types), Default: character(0)}

\item{population_tb}{Population (a tibble), Default: NULL}

\item{region_short_long_chr}{Region short long (a character vector), Default: character(0)}

\item{subdivision_1L_chr}{Subdivision (a character vector of length one), Default: character(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'custom'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'attribute'}

\item{year_1L_chr}{Year (a character vector of length one), Default: character(0)}

\item{...}{Additional arguments}
}
\value{
Object (an output object of multiple potential types)
}
\description{
manufacture method applied to AusLookup
}
