% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_remove.R
\name{remove_empty_lists}
\alias{remove_empty_lists}
\title{Remove empty lists}
\usage{
remove_empty_lists(tbs_ls)
}
\arguments{
\item{tbs_ls}{Tibbles (a list)}
}
\value{
Tibbles (a list)
}
\description{
remove_empty_lists() is a Remove function that edits an object, removing a specified element or elements. Specifically, this function implements an algorithm to remove empty lists. Function argument tbs_ls specifies the object to be updated. The function returns Tibbles (a list).
}
\keyword{internal}
