% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_remove.R
\name{remove_feature_by_prefix_sf}
\alias{remove_feature_by_prefix_sf}
\title{Remove feature by prefix simple features object}
\usage{
remove_feature_by_prefix_sf(data_sf, var_name_1L_chr, prefix_str)
}
\arguments{
\item{data_sf}{Data (a simple features object)}

\item{var_name_1L_chr}{Variable name (a character vector of length one)}

\item{prefix_str}{PARAM_DESCRIPTION}
}
\value{
Data (a simple features object)
}
\description{
remove_feature_by_prefix_sf() is a Remove function that edits an object, removing a specified element or elements. Specifically, this function implements an algorithm to remove feature by prefix simple features object. Function argument data_sf specifies the object to be updated. Argument var_name_1L_chr provides the object to be updated. The function returns Data (a simple features object).
}
\keyword{internal}
