% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_remove.R
\name{remove_feature_no_stat_sf}
\alias{remove_feature_no_stat_sf}
\title{Remove feature no statistic simple features object}
\usage{
remove_feature_no_stat_sf(no_stat_sf, feature_pfx_1L_chr, uids_chr)
}
\arguments{
\item{no_stat_sf}{No statistic (a simple features object)}

\item{feature_pfx_1L_chr}{Feature prefix (a character vector of length one)}

\item{uids_chr}{Unique identifiers (a character vector)}
}
\value{
No statistic (a simple features object)
}
\description{
remove_feature_no_stat_sf() is a Remove function that edits an object, removing a specified element or elements. Specifically, this function implements an algorithm to remove feature no statistic simple features object. Function argument no_stat_sf specifies the object to be updated. Argument feature_pfx_1L_chr provides the object to be updated. The function returns No statistic (a simple features object).
}
\keyword{internal}
