% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_rename.R
\name{rename_by_index}
\alias{rename_by_index}
\title{Rename by index}
\usage{
rename_by_index(df, index_1L_int, new_name_1L_chr)
}
\arguments{
\item{df}{Data.frame (a data.frame)}

\item{index_1L_int}{Index (an integer vector of length one)}

\item{new_name_1L_chr}{New name (a character vector of length one)}
}
\value{
Renamed (a data.frame)
}
\description{
rename_by_index() is a Rename function that renames elements of an object based on a pre-speccified schema. Specifically, this function implements an algorithm to rename by index. The function returns Renamed (a data.frame).
}
\keyword{internal}
