% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_seifa_deciles_tb}
\alias{transform_seifa_deciles_tb}
\title{Transform Socio-Economic Indices for Areas deciles tibble}
\usage{
transform_seifa_deciles_tb(
  seifa_deciles_tb,
  var_nms_1_chr,
  var_nms_2_chr,
  area_unit_1L_chr
)
}
\arguments{
\item{seifa_deciles_tb}{Socio-Economic Indices for Areas deciles (a tibble)}

\item{var_nms_1_chr}{Variable names 1 (a character vector)}

\item{var_nms_2_chr}{Variable names 2 (a character vector)}

\item{area_unit_1L_chr}{Area unit (a character vector of length one)}
}
\value{
Socio-Economic Indices for Areas deciles (a tibble)
}
\description{
transform_seifa_deciles_tb() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform socio-economic indices for areas deciles tibble. Function argument seifa_deciles_tb specifies the object to be updated. Argument var_nms_1_chr provides the object to be updated. The function returns Socio-Economic Indices for Areas deciles (a tibble).
}
\keyword{internal}
