% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_area_names}
\alias{update_area_names}
\title{Update area names}
\usage{
update_area_names(
  area_dss_by_year_ls,
  area_1L_chr,
  altv_names_sf,
  state_nms_chr,
  state_short_and_long_nms_chr,
  remove_symbol_1L_lgl = TRUE
)
}
\arguments{
\item{area_dss_by_year_ls}{Area datasets by year (a list)}

\item{area_1L_chr}{Area (a character vector of length one)}

\item{altv_names_sf}{Alternative names (a simple features object)}

\item{state_nms_chr}{State names (a character vector)}

\item{state_short_and_long_nms_chr}{State short and long names (a character vector)}

\item{remove_symbol_1L_lgl}{Remove symbol (a logical vector of length one), Default: TRUE}
}
\value{
Updated area datasets by year (a list)
}
\description{
update_area_names() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update area names. Function argument area_dss_by_year_ls specifies the object to be updated. Argument area_1L_chr provides the object to be updated. The function returns Updated area datasets by year (a list).
}
\keyword{internal}
