% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aus-package.R
\docType{package}
\name{aus-package}
\alias{aus}
\alias{aus-package}
\title{aus: Auto Underwriting System}
\description{
AUS development and validation.
}
\author{
\strong{Maintainer}: Seokhoon Joo \email{seokhoonj@gmail.com}

}
\keyword{internal}
